+++ core/templatetags/content_tags.py
@@ -6,14 +6,14 @@ register = template.Library()
 
 
 @register.filter
-def format_timedelta(timedelta):
+def format_timedelta(timedelta, pluralize=False):
 
     if isinstance(timedelta, datetime.timedelta):
         # round up to next minute
         rounded_mins = math.ceil(timedelta.total_seconds() / 60)
         hours, mins = divmod(rounded_mins, 60)
-        hours_plural = 's' if hours > 1 else ''
-        mins_plural = 's' if mins > 1 else ''
+        hours_plural = 's' if hours > 1 and pluralize else ''
+        mins_plural = 's' if mins > 1 and pluralize else ''
         hours_str = f'{hours} hour{hours_plural}' if hours else ''
         mins_str = f'{mins} min{mins_plural}' if mins or not hours else ''
         return f'{hours_str} {mins_str}'.strip()