+++ react-components/src/views/LoginModal/Step1.jsx
@@ -3,20 +3,24 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Services from '@src/Services'
-import Field from '@src/components/Field'
+import Field from '@src/components/Fields/Field'
 import SocialLoginButtons from '@src/components/SocialLoginButtons'
 import ErrorList from '@src/components/ErrorList'
-
 import './stylesheets/Step1.scss'
 
-
-export default function Step1(props){
+export default function Step1(props) {
   return (
-    <div className='great-mvp-signup-wizard-step-1'>
+    <div className="great-signup-wizard-step-1">
       <h2 className="h-xl m-t-l">Log in</h2>
-      <form onSubmit={event => {event.preventDefault(); props.handleSubmit() }}>
+      <form
+        onSubmit={(event) => {
+          event.preventDefault()
+          props.handleSubmit()
+        }}
+      >
         <ErrorList errors={props.errors.__all__ || []} className="m-b-s" />
         <Field
+          id="id_email"
           type="text"
           placeholder="Email address"
           name="email"
@@ -27,6 +31,7 @@ export default function Step1(props){
           errors={props.errors.email || []}
         />
         <Field
+          id="id_password"
           type="password"
           placeholder="Password"
           name="password"
@@ -38,23 +43,24 @@ export default function Step1(props){
         <input
           type="submit"
           value="Log in"
-          className="great-mvp-wizard-step-submit great-mvp-wizard-step-button"
+          className="great-wizard-step-submit great-wizard-step-button"
           disabled={props.disabled}
         />
-        <p><a href={Services.config.passwordResetUrl}>Forgotten password?</a></p>
-      </form>          
+        <p>
+          <a href={Services.config.passwordResetUrl}>Forgotten password?</a>
+        </p>
+      </form>
 
-      <div className='great-mvp-vertical-separator'>
-        <hr/>
+      <div className="great-vertical-separator">
+        <hr />
         <span>or</span>
-        <hr/>
+        <hr />
       </div>
 
-      <SocialLoginButtons
-        linkedinUrl={props.linkedinLoginUrl}
-        googleUrl={props.googleLoginUrl}
-      />
-      <p>Do not have an account? <a href={Services.config.signupUrl}>Sign up</a></p>
+      <SocialLoginButtons linkedinUrl={props.linkedinLoginUrl} googleUrl={props.googleLoginUrl} />
+      <p>
+        Do not have an account? <a href={Services.config.signupUrl}>Sign up</a>
+      </p>
     </div>
   )
 }
@@ -66,12 +72,12 @@ Step1.propTypes = {
   handleSubmit: PropTypes.func.isRequired,
   handleEmailChange: PropTypes.func.isRequired,
   password: PropTypes.string,
-  email: PropTypes.string,
+  email: PropTypes.string
 }
 
 Step1.defaultProps = {
   disabled: false,
   errors: {},
   password: '',
-  email: '',
+  email: ''
 }