+++ react-components/src/views/Tour/Container.jsx
@@ -3,21 +3,19 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import ReactDOM from 'react-dom'
 
-import { useCookies } from 'react-cookie';
+import { useCookies } from 'react-cookie'
 import ReactModal from 'react-modal'
 
 import Component from './Component'
 
-
-export function Container(props){
-  const [cookies, setCookie] = useCookies([props.disableTourCookieName]); 
+export function Container(props) {
+  const [cookies, setCookie] = useCookies([props.disableTourCookieName])
   const [isOpenModal, setIsOpenModal] = React.useState(cookies[props.disableTourCookieName] !== 'true')
   const [isOpenTour, setIsOpenTour] = React.useState()
 
   function handleSkipTour(error) {
     setIsOpenModal(false)
     setIsOpenTour(false)
-    setCookie(props.disableTourCookieName, 'true');
   }
 
   function handleStartTour(nextStep) {
@@ -27,7 +25,6 @@ export function Container(props){
 
   function handleTourClose() {
     setIsOpenTour(false)
-    setCookie(props.disableTourCookieName, 'true');
   }
 
   return (
@@ -46,12 +43,12 @@ export function Container(props){
 }
 
 Container.propTypes = {
-  isOpen: PropTypes.bool,
+  isOpen: PropTypes.bool
 }
 
 Container.defaultProps = {
   isOpenTour: false,
-  isOpenModal: false,
+  isOpenModal: false
 }
 
 export default function createModal({ element, ...params }) {