+++ react-components/src/views/Tour/Modal.jsx
@@ -5,12 +5,10 @@ import ReactModal from 'react-modal'
 
 import Services from '@src/Services'
 
-
 import './stylesheets/Tour.scss'
 
-
 const modalStyles = {
-  content : {
+  content: {
     background: '#f5f2ed',
     bottom: 'auto',
     left: '50%',
@@ -21,48 +19,50 @@ const modalStyles = {
     transform: 'translate(-50%, -50%)',
     borderRadius: 10,
     padding: 0,
-    width: 490,
+    width: 490
   },
   overlay: {
     zIndex: 1000,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
+    backgroundColor: 'rgba(0, 0, 0, 0.5)'
+  }
 }
 
-
-export default function Modal(props){
-
+export default function Modal(props) {
   return (
-    <ReactModal
-      isOpen={props.isOpen}
-      style={modalStyles}
-      onRequestClose={props.handleSkip}
-      contentLabel="Modal"
-    >
-      <div className="great-mvp-signup-wizard-step-1 p-s" id="page-tour-modal-step-1">
-        <h2 className="h-m">{props.title}</h2>
-        <p>{props.body}</p>
-        <input
-          type="submit"
-          value={props.buttonText}
-          className="great-mvp-tour-button p-v-xxs p-h-xs"
-          id="page-tour-submit"
-          onClick={event => { event.preventDefault(); props.handleStart()}}
-        />
-          <p className="m-t-xxs">
-            <a
-              href="#"
-              id="page-tour-skip"
-              onClick={event => { event.preventDefault(); props.handleSkip()}}
-            >Skip walk-through</a>
-          </p>
+    <ReactModal isOpen={props.isOpen} style={modalStyles} onRequestClose={props.handleSkip} contentLabel="Modal">
+      <div className="great-tour-modal great-signup-wizard-step-1" id="page-tour-modal-step-1">
+        <img src="/static/images/learning-modal.png" alt="" />
+        <h2 className="great-tour-heading m-t-s">{props.title}</h2>
+        <p className="great-tour-text">{props.body}</p>
+        <div className="great-tour-actions">
+          <button
+            id="page-tour-submit"
+            className="button button--primary button--large"
+            onClick={(event) => {
+              event.preventDefault()
+              props.handleStart()
+            }}
+          >
+            {props.buttonText}
+          </button>
+          <a
+            id="page-tour-skip"
+            className="link"
+            href="#"
+            onClick={(event) => {
+              event.preventDefault()
+              props.handleSkip()
+            }}
+          >
+            Skip walk-through
+          </a>
+        </div>
       </div>
     </ReactModal>
   )
- 
 }
 
 Modal.propTypes = {
   handleSkip: PropTypes.func.isRequired,
-  handleStart: PropTypes.func.isRequired,
+  handleStart: PropTypes.func.isRequired
 }