+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
+- AA-129 - wagtail import export for Beta environment
+- GP2-270 - Homepage editable
+
 ## Pre-release
 
+- GP2-218 - Image and media surfacing from CMS modules
+- No ticket - Wagtail upgrade to 2.9.3
+
 ### Fixed bugs
+
+- GP2-280 - removal of product selector from dashboard
+- GP-310 - Format time as min rather than mins
+- GP2-247 - image resizing module image
+- GP2-246 - wrong url for learning header link
 - GP-96 - HS Codes saved in DirectoryAPI
+- GP2-176 - Navigation to the next section
+- GP-96 - HS Codes saved in DirectoryAPI
+- GP2-113 - Add objectives list component
 - MVP-431 - Fix percentage amounts in country data component
 - MVP-428 - Fix country data component formatting
 - MVP-402 - Fix Localisation
@@ -13,9 +27,18 @@
 - MVP-443 - Use non-GREAT stlye 404 and 500 pages
 - MVP-579 - create export plan cms page
 - GP2-67 - fix backbutton upload logo
+- GP2-170 - business performance
 
 ### Implemented enhancements
 
+- GP2-159 - learning onboarding - tooltips
+- GP2-160 - learning onboarding - skip walkthrough
+- GP2-156 - learning onboarding - welcome message
+- GP2-307 - Add new form field to the target markets research page
+- GP2-207 - Welcome back message
+- GP2-197 - Export plan navigation to the next section
+- GP2-280 - Removal of product sector selector modal
+- GP2-192 - Marketing approach page structure
 - GP2-23 - Modules - not started
 - GP2-31 - Count of topics and lessons in a module
 - GP2-73 - Module landing page - topics & lessons
@@ -118,3 +141,4 @@
 - GP2-120 - Fix numbers on curated list page
 - GP2-163 - Show estimated reading time for lesson
 - no-ticket - Migrate to codecov remove coveralls
+- no-ticket - Bump django version to 2.2.14